// src/modules/referrals/utils/referrals.enum.ts
export enum REFERRAL_STATUS {
  PENDING = "Pending",
  INVITED = "Invited",
  QUALIFIED = "Qualified",
  EXPIRED = "Expired",
}
export enum REWARD_STATUS {
  EARNED = "Earned",
  PENDING = "Pending",
  REDEEMED = "Redeemed",
}

export enum REWARD_TYPE {
  CREDIT = "CREDIT",
  POINTS = "POINTS",
}

export enum DURATION {
  MONTHLY = "monthly",
  YEARLY = "yearly",
  WEEKLY = "weekly",
}

export enum REFERRAL_ROLE {
  REFEREE = "referee",
  REFERER = "referer",
}

export enum REFERRAL_REWARD_TRIGGER {
  SIGN_UP = "sign_up",
}